[BE] mock payment mechanism

+++ PerRead.Backend/Models/FrontEnd/FEUserPreview.cs
@@ -8,6 +8,6 @@
 
         public string ProfileImageUri { get; set; }
 
-        public int ReadingTokens { get; set; }
+        public long ReadingTokens { get; set; }
     }
 }
